+++ backend/src/main/kotlin/com/back/back9/domain/user/dto/UserLoginDto.kt
@@ -1,0 +1,17 @@
+package com.back.back9.domain.user.dto
+
+import jakarta.validation.constraints.NotBlank
+import jakarta.validation.constraints.Size
+
+data class UserLoginReqBody(
+    @field:NotBlank @field:Size(min = 2, max = 30)
+    val userLoginId: String,
+    @field:NotBlank @field:Size(min = 2, max = 30)
+    val password: String
+)
+
+data class UserLoginResBody(
+    val item: UserDto,
+    val apiKey: String,
+    val accessToken: String
+)
