+++ backend/src/main/kotlin/com/back/back9/domain/user/dto/UserRegisterDto.kt
@@ -1,15 +1,16 @@
 package com.back.back9.domain.user.dto
 
 import jakarta.validation.constraints.NotBlank
+import jakarta.validation.constraints.Size
 
 data class UserRegisterDto(
-    @field:NotBlank
+    @field:NotBlank @field:Size(min = 2, max = 30)
     val userLoginId: String,
-    @field:NotBlank
+    @field:NotBlank @field:Size(min = 2, max = 30)
     val username: String,
-    @field:NotBlank
+    @field:NotBlank @field:Size(min = 2, max = 30)
     val password: String,
-    @field:NotBlank
+    @field:NotBlank @field:Size(min = 2, max = 30)
     val confirmPassword: String
 ) {
     init {
