+++ backend/src/test/kotlin/com/back/BackendApplicationTests.kt
@@ -1,0 +1,12 @@
+package com.back
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class BackendApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+}
